+++ modules/apps/headless/headless-collaboration/headless-collaboration-impl/src/main/java/com/liferay/headless/collaboration/internal/resource/v1_0/BlogPostingResourceImpl.java
@@ -25,14 +25,11 @@ import com.liferay.document.library.kernel.service.DLAppService;
 import com.liferay.document.library.util.DLURLHelper;
 import com.liferay.headless.collaboration.dto.v1_0.BlogPosting;
 import com.liferay.headless.collaboration.dto.v1_0.Image;
-import com.liferay.headless.collaboration.dto.v1_0.Rating;
 import com.liferay.headless.collaboration.dto.v1_0.TaxonomyCategory;
 import com.liferay.headless.collaboration.internal.dto.v1_0.util.AggregateRatingUtil;
 import com.liferay.headless.collaboration.internal.dto.v1_0.util.CreatorUtil;
-import com.liferay.headless.collaboration.internal.dto.v1_0.util.RatingUtil;
 import com.liferay.headless.collaboration.internal.odata.entity.v1_0.BlogPostingEntityModel;
 import com.liferay.headless.collaboration.resource.v1_0.BlogPostingResource;
-import com.liferay.headless.common.spi.resource.SPIRatingResource;
 import com.liferay.headless.common.spi.service.context.ServiceContextUtil;
 import com.liferay.portal.kernel.comment.CommentManager;
 import com.liferay.portal.kernel.model.User;
@@ -91,16 +88,6 @@ public class BlogPostingResourceImpl
 	}
 
 	@Override
-	public Page<Rating> getBlogPostingsRatingsPage(Long blogPostingId)
-		throws Exception {
-
-		SPIRatingResource<Rating> spiRatingResource = _getSPIRatingResource();
-
-		return spiRatingResource.getRatingsPage(
-			BlogsEntry.class.getName(), blogPostingId);
-	}
-
-	@Override
 	public Page<BlogPosting> getContentSpaceBlogPostingsPage(
 			Long contentSpaceId, Filter filter, Pagination pagination,
 			Sort[] sorts)
@@ -127,16 +114,6 @@ public class BlogPostingResourceImpl
 	@Override
 	public EntityModel getEntityModel(MultivaluedMap multivaluedMap) {
 		return _entityModel;
-	}
-
-	@Override
-	public Rating postBlogPostingRating(Long blogPostingId, Rating rating)
-		throws Exception {
-
-		SPIRatingResource<Rating> spiRatingResource = _getSPIRatingResource();
-
-		return spiRatingResource.postRating(
-			BlogsEntry.class.getName(), blogPostingId, GetterUtil.getDouble(rating.getRatingValue()));
 	}
 
 	@Override
@@ -259,14 +236,6 @@ public class BlogPostingResourceImpl
 			throw new BadRequestException(
 				"Unable to get file entry " + imageId, e);
 		}
-	}
-
-	private SPIRatingResource<Rating> _getSPIRatingResource() {
-		return new SPIRatingResource<>(
-			_ratingsEntryLocalService,
-			ratingsEntry -> RatingUtil.toRating(
-				_portal, ratingsEntry, _userLocalService),
-			_user);
 	}
 
 	private BlogPosting _toBlogPosting(BlogsEntry blogsEntry) throws Exception {
