+++ modules/apps/headless/headless-collaboration/headless-collaboration-impl/src/main/java/com/liferay/headless/collaboration/internal/resource/v1_0/RatingResourceImpl.java
@@ -14,14 +14,20 @@
 
 package com.liferay.headless.collaboration.internal.resource.v1_0;
 
+import com.liferay.blogs.model.BlogsEntry;
 import com.liferay.headless.collaboration.dto.v1_0.Rating;
 import com.liferay.headless.collaboration.internal.dto.v1_0.util.RatingUtil;
 import com.liferay.headless.collaboration.resource.v1_0.RatingResource;
 import com.liferay.headless.common.spi.resource.SPIRatingResource;
+import com.liferay.knowledge.base.model.KBArticle;
+import com.liferay.message.boards.model.MBMessage;
+import com.liferay.message.boards.model.MBThread;
+import com.liferay.message.boards.service.MBThreadLocalService;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.service.UserLocalService;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.Portal;
+import com.liferay.portal.vulcan.pagination.Page;
 import com.liferay.ratings.kernel.service.RatingsEntryLocalService;
 
 import javax.ws.rs.core.Context;
@@ -47,10 +53,106 @@ public class RatingResourceImpl extends BaseRatingResourceImpl {
 	}
 
 	@Override
+	public Page<Rating> getBlogPostingRatingsPage(Long blogPostingId)
+		throws Exception {
+
+		SPIRatingResource<Rating> spiRatingResource = _getSPIRatingResource();
+
+		return spiRatingResource.getRatingsPage(
+			BlogsEntry.class.getName(), blogPostingId);
+	}
+
+	@Override
+	public Page<Rating> getKnowledgeBaseArticleRatingsPage(
+			Long knowledgeBaseArticleId)
+		throws Exception {
+
+		SPIRatingResource<Rating> spiRatingResource = _getSPIRatingResource();
+
+		return spiRatingResource.getRatingsPage(
+			KBArticle.class.getName(), knowledgeBaseArticleId);
+	}
+
+	@Override
+	public Page<Rating> getMessageBoardMessageRatingsPage(
+			Long messageBoardMessageId)
+		throws Exception {
+
+		SPIRatingResource<Rating> spiRatingResource = _getSPIRatingResource();
+
+		return spiRatingResource.getRatingsPage(
+			MBMessage.class.getName(), messageBoardMessageId);
+	}
+
+	@Override
+	public Page<Rating> getMessageBoardThreadRatingsPage(
+			Long messageBoardThreadId)
+		throws Exception {
+
+		SPIRatingResource<Rating> spiRatingResource = _getSPIRatingResource();
+
+		MBThread mbThread = _mbThreadLocalService.getMBThread(
+			messageBoardThreadId);
+
+		return spiRatingResource.getRatingsPage(
+			MBMessage.class.getName(), mbThread.getRootMessageId());
+	}
+
+	@Override
 	public Rating getRating(Long ratingId) throws Exception {
 		SPIRatingResource<Rating> spiRatingResource = _getSPIRatingResource();
 
 		return spiRatingResource.getRating(ratingId);
+	}
+
+	@Override
+	public Rating postBlogPostingRating(Long blogPostingId, Rating rating)
+		throws Exception {
+
+		SPIRatingResource<Rating> spiRatingResource = _getSPIRatingResource();
+
+		return spiRatingResource.postRating(
+			BlogsEntry.class.getName(), blogPostingId,
+			GetterUtil.getDouble(rating.getRatingValue()));
+	}
+
+	@Override
+	public Rating postKnowledgeBaseArticleRating(
+			Long knowledgeBaseArticleId, Rating rating)
+		throws Exception {
+
+		SPIRatingResource<Rating> spiRatingResource = _getSPIRatingResource();
+
+		return spiRatingResource.postRating(
+			KBArticle.class.getName(), knowledgeBaseArticleId,
+			GetterUtil.getDouble(rating.getRatingValue()));
+	}
+
+	@Override
+	public Rating postMessageBoardMessageRating(
+			Long messageBoardMessageId, Rating rating)
+		throws Exception {
+
+		SPIRatingResource<Rating> spiRatingResource = _getSPIRatingResource();
+
+		return spiRatingResource.postRating(
+			MBMessage.class.getName(), messageBoardMessageId,
+			GetterUtil.getDouble(rating.getRatingValue()));
+	}
+
+	@Override
+	public Rating postMessageBoardThreadRating(
+			Long messageBoardThreadId, Rating rating)
+		throws Exception {
+
+		SPIRatingResource<Rating> spiRatingResource = _getSPIRatingResource();
+
+		MBThread mbThread = _mbThreadLocalService.getThread(
+			messageBoardThreadId);
+
+		return spiRatingResource.postRating(
+			MBMessage.class.getName(), mbThread.getRootMessageId(),
+			GetterUtil.getDouble(rating.getRatingValue()));
 	}
 
 	@Override
@@ -68,6 +170,9 @@ public class RatingResourceImpl extends BaseRatingResourceImpl {
 				_portal, ratingsEntry, _userLocalService),
 			_user);
 	}
+
+	@Reference
+	private MBThreadLocalService _mbThreadLocalService;
 
 	@Reference
 	private Portal _portal;
